SalesMicroservice/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-jose
httpx